// types/job/interview.ts
export interface Interview {
  id: string;
  jobPostId: string;
  memberId: string;
  corporateId: string;
  roomId?: string | undefined;
  interviewDate?: string | undefined;
  status: string;
  failureReason?: string | undefined;
  createdAt: string;
  updatedAt: string;
  jobPost?: JobPost | undefined;
  reviews: Review[];
}

export interface JobPost {
  id: string;
  title?: string | undefined;
  corporateId: string;
}

export interface Review {
  id: string;
  interviewId: string;
  reviewerId: string;
  rating?: number | undefined;
  comment?: string | undefined;
  createdAt: string;
}

/**
 * ---------------------
 * CRUD Responses
 * ---------------------
 */
export interface InterviewResponse {
  success: boolean;
  message?: string | undefined;
  data?: Interview | undefined;
}

export interface ListInterviewResponse {
  message?: string | undefined;
  interview: Interview | undefined;
}

export interface AllInterviewsWithMetaResponse {
  data: Interview[];
  meta: PaginationMeta | undefined;
}

export interface DeleteInterviewResponse {
  success: boolean;
  message: string;
}